Add results for dim0 optimization
In the neighborhoods iterable interval iterator, one can optimize the
code for accessing dimension[0] and minimum[0] only once per iterator
lifetime and storing the value in a constant.

This optimization brings a speedup of 2-4% and does not reduce
readability to much, since dimension0 is pretty similar to the previous
name.